convert explorer to new API
