+++ client/src/explorer_common/explorer_components.js
@@ -1,3 +1,4 @@
+import "./explorer-styles.scss";
 import "./explorer-components.scss";
 
 import classNames from "classnames";
